added easy port mapping function

+++ servers/upnp/types/portMapping.go
@@ -3,7 +3,7 @@ package types
 type PortMapping struct {
 	ExternalPort uint16
 	Protocol     string
-	InternalPort int
+	InternalPort uint16
 	InternalIP   string
 	Enabled      bool
 	Description  string
